requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.2
yfinance==0.2.36
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
